requirements.txt
Flask-RESTful==0.3.6
Flask==0.12.2
Flask-SQLAlchemy==2.1
mysqlclient==1.3.12
flask-cors==3.0.3
python-jose==2.0.2